함수명을 변경합니다
