--- docs/src/pages/docs/api/getHangulSimilarity.ko.mdx
@@ -1,26 +1,0 @@
-import { GetHangulSimilarityDemo } from '@/components/demo/get-hangul-similarity-demo';
-
-# getHangulSimilarity
-
-두 한글 문자열의 유사도를 계산하여 0에서 1 사이의 값을 반환합니다. 이 함수는 es-hangul 라이브러리의 disassembleHangul 메서드를 사용하여 한글 문자열을 자모 단위로 분해하고, Levenshtein 거리 알고리즘을 이용해 문자열 간의 최소 편집 거리를 계산한 후 이를 바탕으로 유사도를 계산합니다.
-
-```tsx
-function getHangulSimilarity(
-  // 비교할 첫 번째 한글 문자열
-  left: string,
-  // 비교할 두 번째 한글 문자열
-  right: string
-): number;
-```
-
-## Examples
-
-```tsx
-getHangulSimilarity('안녕하세요', '안녕하세요'); // 1
-getHangulSimilarity('안녕하세요', '안녕하새요'); // 약 0.916
-getHangulSimilarity('동해물과', '동해붇ㄹㅁ과'); // 약0.769
-getHangulSimilarity('대한민국', '대한밍굮'); // 약 0.818
-getHangulSimilarity('가나다라마바사', '나다라마바사가'); // 약 0.714
-```
-
-<GetHangulSimilarityDemo />
