+++ docs/src/pages/docs/api/getSimilarity.en.mdx
@@ -1,0 +1,25 @@
+import { GetSimilarityDemo } from '@/components/demo/get-similarity-demo';
+
+# getSimilarity
+
+Calculates the similarity between two Korean strings and returns a value between 0 and 1. This function uses the disassemble method from the es-hangul library to break down the Korean strings into their component characters (jamo) and then calculates the minimum edit distance between the strings using the Levenshtein distance algorithm. The similarity is then calculated based on this distance.
+
+```tsx
+function getSimilarity(
+  // First Korean string to compare
+  left: string,
+  // Second Korean string to compare
+  right: string
+): number;
+
+## Examples
+
+```tsx
+getSimilarity('안녕하세요', '안녕하세요'); // 1
+getSimilarity('안녕하세요', '안녕하새요'); // 약 0.916
+getSimilarity('동해물과', '동해붇ㄹㅁ과'); // 약 0.769
+getSimilarity('대한민국', '대한밍굮'); // 약0.818
+getSimilarity('가나다라마바사', '나다라마바사가'); // 약0.714
+```
+
+<GetSimilarityDemo />
