+++ src/index.ts
@@ -8,4 +8,4 @@ export * from './hangulIncludes';
 export * from './josa';
 export * from './removeLastHangulCharacter';
 export * from './utils';
-export * from './getHangulSimilarity';
+export * from './getSimilarity';
